corrido um erro

+++ service/util/doc_swagger.py
@@ -3,4 +3,4 @@ from service.restplus import api
 
 INPUT_MAIN_SERVICE = api.model(
   'input_main_service', {
-    'valorM': fields.List(fields.Integer(), required=True, description= "Valor em metros")})
+    'valorM': fields.List(fields.Float(), required=True, description= "Valor em metros")})
